correct?

+++ exam4/microshell.c
@@ -5,9 +5,9 @@
 #include <sys/errno.h>
 #include <stdio.h>
 
-#define STDIN		0
-#define STDOUT		1
-#define STDERR		2
+#define STDIN 0
+#define STDOUT 1
+#define STDERR 2
 
 char **g_envp = NULL;
 
@@ -17,31 +17,33 @@ t_token *g_tokens = NULL;
 size_t g_program_count = 1;
 t_program *g_programs = NULL;
 
-int terminate(char *msg, int ret)
+int terminate(int ret)
 {
-	// if (msg)
+	if (g_programs)
+	{
+		for (size_t i = 0; i < g_program_count; ++i)
+			free(g_programs[i].args);
+		free(g_programs);
+		g_programs = NULL;
+	}
+	if (g_tokens)
+	{
+		free(g_tokens);
+		g_tokens = NULL;
+	}
+
+	// while (1)
 	// {
-	// 	ft_putstr(FD_ERR, msg);
-	// 	ft_putstr(FD_ERR, "\n");
 	// }
-
-	for (size_t i = 0; i < g_program_count; ++i)
-		free(g_programs[i].args);
-
-	free(msg);
-	free(g_tokens);
-	free(g_programs);
-
 	exit(ret);
 
 	return (0);
 }
 
-
 void exit_fatal(void)
 {
 	write(STDERR, "error: fatal\n", ft_strlen("error: fatal\n"));
-	terminate(NULL, 1);
+	terminate(1);
 	exit(EXIT_FAILURE);
 }
 
@@ -50,15 +52,15 @@ void exit_execve(char *str)
 	write(STDERR, "error: cannot execute ", ft_strlen("error: cannot execute "));
 	write(STDERR, str, ft_strlen(str));
 	write(STDERR, "\n", 1);
-	exit(EXIT_FAILURE);
+	// terminate(1);
+	// exit(EXIT_FAILURE);
 }
 
 int exit_cd_1()
 {
 	write(STDERR, "error: cd: bad arguments\n", ft_strlen("error: cd: bad arguments\n"));
-	//return (EXIT_FAILURE);
-	terminate(NULL, 1);
-	return(EXIT_FAILURE);
+	terminate(1);
+	return (EXIT_FAILURE);
 }
 
 int exit_cd_2(char *str)
@@ -66,72 +68,50 @@ int exit_cd_2(char *str)
 	write(STDERR, "error: cd: cannot change directory to ", ft_strlen("error: cd: cannot change directory to "));
 	write(STDERR, str, ft_strlen(str));
 	write(STDERR, "\n", 1);
+	terminate(1);
 	return (EXIT_FAILURE);
 }
 
-
-int terminate_errno(char *origin)
-{
-	ft_putstr(FD_ERR, "microshell: ");
-	if (origin)
-	{
-		ft_putstr(FD_ERR, origin);
-		ft_putstr(FD_ERR, ": ");
-	}
-	ft_putstr(FD_ERR, strerror(errno));
-	ft_putstr(FD_ERR, "\n");
-
-	return (terminate(NULL, true));
-}
-
-void			control_c(int sig)
-{
-
-
-	(void)sig;
-	for (size_t i = 0; i < g_program_count; ++i)
-	{
-		if(g_programs[i].pid != 0)
-		{
-			kill(g_programs[i].pid, sig);
-			printf("KILL\n");
-		}
-	}
-	// write(1, "\n", 1);
-	terminate(NULL, 130);
-	
-}
-
-void			control_quit(int sig)
+void control_c(int sig)
 {
 
 	(void)sig;
-	int is_kill = 0;
-	for (size_t i = 0; i < g_program_count; ++i)
+	if (g_programs)
 	{
-		if(g_programs[i].pid != 0)
+		for (size_t i = 0; i < g_program_count; ++i)
 		{
-			kill(g_programs[i].pid, sig);
-			is_kill = 1;
+			if (g_programs[i].pid != 0)
+			{
+				kill(g_programs[i].pid, sig);
+				printf("KILL\n");
+			}
 		}
 	}
-
-	if (is_kill)
-	{
-		write(1, "\b\b  \b\b", 6);
-		terminate(NULL, 127);
-	}
-	else
-	{
-		write(1, "Quit\n", 5);
-		terminate(NULL, 131);
-	}
-	
+	terminate(130);
 }
+
+void control_quit(int sig)
+{
+
+	(void)sig;
+	if (g_programs)
+	{
+		for (size_t i = 0; i < g_program_count; ++i)
+		{
+			if (g_programs[i].pid != 0)
+			{
+				kill(g_programs[i].pid, sig);
+			}
+		}
+	}
+	terminate(131);
+}
+
 int main(int argc, char **argv, char **envp)
 {
 	g_envp = envp;
-
+	int ret = 0;
+	int status;
 	if (signal(SIGINT, &control_c) == SIG_ERR)
 	{
 		exit_fatal();
@@ -148,13 +128,11 @@ int main(int argc, char **argv, char **envp)
 
 	g_tokens = malloc_zeros((g_token_count = argc - 1) * sizeof(t_token));
 
-//	printf("g_token_count: %zu\n", g_token_count);
-	/* Arguments to tokens. */
 	for (int i = 1; i < argc; ++i)
 	{
 		t_token *tok = &(g_tokens[i - 1]);
 		char *curr = tok->data = argv[i];
-//		printf("i: %d / data : %s\n", i, tok->data);
+
 		tok->end = i == argc - 1;
 
 		if (strcmp(";", curr) == 0)
@@ -170,21 +148,20 @@ int main(int argc, char **argv, char **envp)
 
 	g_programs = malloc_zeros(g_program_count * sizeof(t_program));
 
-	/* Tokens to 'program' struct. */
 	{
 		size_t j = 0;
 
 		for (size_t i = 0; i < g_program_count; ++i)
 		{
 			t_program *pr = &(g_programs[i]);
-		//	printf("j before : %zu \n", j);
+
 			size_t start = j;
-		//	printf("j after : %zu \n", j);
+
 			t_token *tok = NULL;
 			while (j < g_token_count)
 			{
 				tok = &(g_tokens[j]);
-		//		printf("j loop : %zu \n", j);
+
 				if (tok->type != TT_STRING)
 				{
 					pr->piped = tok->type == TT_PIPE;
@@ -194,41 +171,25 @@ int main(int argc, char **argv, char **envp)
 				j++;
 			}
 
-		//	printf("j : %zu \n", j);
-		///	printf("start : %zu \n", start);
 			pr->count = j - start;
-		//	printf("count : %zu\n", pr->count);
+
 			if (pr->count == 0 && (!tok || tok->type == TT_SEMICOLON))
 			{
 				j++;
-		//		printf("j in the loop : %zu\n", j);
 				continue;
 			}
-		//	printf("count : %zu\n", pr->count);
+
 			pr->args = malloc_zeros((pr->count + 1) * sizeof(char *));
 
 			for (size_t k = 0; k < pr->count; ++k)
 			{
 				pr->args[k] = g_tokens[start + k].data;
-		//		printf("arg: %s\n", pr->args[k]);
 			}
 
 			pr->path = pr->args[0];
-			 j++;
+			j++;
 		}
 	}
-
-	//	/* Dump. */
-	//	for (size_t i = 0; i < g_program_count; ++i)
-	//	{
-	//		t_program *pr = &(g_programs[i]);
-	//
-	//		printf("\n\npath = %s\n", pr->path);
-	//		printf("piped = %s\n", pr->piped ? "true" : "false");
-	//		printf("semicoloned = %s\n", pr->semicoloned ? "true" : "false");
-	//		for (size_t j = 0; pr->args[j]; ++j)
-	//			printf("arg[%zu] = %s\n", j, pr->args[j]);
-	//	}
 
 	int fd_in = 0;
 
@@ -236,123 +197,95 @@ int main(int argc, char **argv, char **envp)
 	for (size_t i = 0; i < g_program_count; ++i)
 	{
 		t_program *pr = &(g_programs[i]);
-		if(pr->args)
+		if (pr->args)
 		{
-			int p[2];
-			pipe(p);
 
-			pid_t pid = fork();
-
-			if (pid == 0)
-			{
-				dup2(fd_in, 0);
-
-				if (pr->piped)
-					dup2(p[1], 1);
-
-				close(p[0]);
-				execve(pr->path, pr->args, envp);
-			}
-			else
-			{
-				if (pr->semicoloned || i == g_program_count - 1)
-				{
-					waitpid(pid, NULL, 0);
-					if (fd_in)
-						close(fd_in);
-					fd_in = 0;
-					close(p[0]);
-				}
-				else
-				{
-					if (fd_in)
-						close(fd_in);
-					fd_in = p[0];	
-				}
-
-				close(p[1]);
-			}
-			
-
-			/*if (strcmp("cd", pr->path) == 0)
+			if (strcmp("cd", pr->path) == 0)
 			{
 				if (pr->count < 2)
 					exit_cd_1();
-				
+
 				ft_putstr(1, pr->args[1]);
 				if (chdir(pr->args[1]))
 					exit_cd_2(pr->args[1]);
-				// else
-				// {
-				// 	char buf[500];
-				// 	getcwd(buf, 500);
-				// 	printf("buf:%s\n", buf);
-				// }
 			}
-			
+
 			else
-			{*/
-				
-			/*
-				if(pipe(pr->pipes))
+			{
+				if (pipe(pr->pipes))
 					exit_fatal();
 
 				if ((pr->pid = fork()) == -1)
 					exit_fatal();
 				else if (pr->pid == 0)
 				{
-					if (dup2(fd_in, FD_IN) < 0)
-						exit_fatal();
-			//		printf("fd in after close : %d\n", fd_in);
+					if (fd_in)
+					{
+						if (dup2(fd_in, FD_IN) < 0)
+							exit_fatal();
+						close(fd_in);
+					}
+
 					if (pr->piped)
 					{
 						if (dup2(pr->pipes[1], FD_OUT) < 0)
 							exit_fatal();
+						close(pr->pipes[1]);
 					}
-					close(pr->pipes[0]);
-					printf("count: %zu\n", pr->count);
-					if (execve(pr->path, pr->args, g_envp) == -1) 
+
+					if (execve(pr->path, pr->args, g_envp) == -1)
+					{
+						//exit(1);
+						close(pr->pipes[0]);
+						
+						//fd_in = pr->pipes[0];
+
 						exit_execve(pr->path);
+						if (g_programs)
+						{
+							for (size_t i = 0; i < g_program_count; ++i)
+								free(g_programs[i].args);
+							free(g_programs);
+							g_programs = NULL;
+						}
+						if (g_tokens)
+						{
+							free(g_tokens);
+							g_tokens = NULL;
+						}
+						exit(1);
+						
+					}
 					exit(EXIT_SUCCESS);
 				}
 				else
 				{
 					if (pr->semicoloned || i == g_program_count - 1)
 					{
-						//				ft_putstr(FD_ERR, strcat(strcat(strdup("waiting: "), pr->path), "\n"));
-						dprintf(2, "----------------===-=-=-=- waiting %d: %s %zu\n", pr->pid, pr->path, i);
-						waitpid(pr->pid, NULL, 0);
-						dprintf(2, "----------------===-=-=-=- finished waiting %d\n", pr->pid);
+						waitpid(pr->pid, &status, 0);
+						if (WIFEXITED(status))
+							ret = WEXITSTATUS(status);
 						close(pr->pipes[0]);
-						if (fd_in)
-							close(fd_in);
+
 						fd_in = 0;
 					}
 					else
 					{
 						fd_in = pr->pipes[0];
-						// printf("fd in : %d\n", fd_in);
 					}
-					//if (i > 0 && g_programs[i-1].piped)
-					//	close(g_programs[i-1].pipes[0]);
+					if (i > 0 && g_programs[i - 1].piped)
+					{
+						close(g_programs[i - 1].pipes[0]);
+					}
 					close(pr->pipes[1]);
-				}*/
-			//}
+				}
+			}
 		}
 		else
 		{
-			break;
+			return (terminate(1));
 		}
-		
 	}
 
-	if (fd_in != 0)
-		close(fd_in);
-
-	/* Just to be sure. */
-	//for (size_t i = 0; i < g_program_count; ++i)
-	//	waitpid(g_programs[i].pid, NULL, 0);
-	// while(1)
-	// {}
-	return (terminate(NULL, false));
+	return (terminate(ret));
 }
